progress

+++ BulletinBoard/BulletinBoardApi/Models/ImageDB.cs
@@ -9,10 +9,10 @@ namespace BulletinBoardApi.Models
         [BsonId(IdGenerator = typeof(ObjectIdGenerator))]
         public ObjectId? _id { get; set; }
         public string Image64 { get; set; }
-        public int X { get; set; }
-        public int Y { get; set; }
-        public int Width { get; set; }
-        public int Height { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
+        public double Width { get; set; }
+        public double Height { get; set; }
         public double Angle { get; set; }
 
     }
